fix up reconf.sh a bit
git-svn-id: svn://svn.valgrind.org/valkyrie/trunk@43 512c2683-75f9-0310-acad-d911283a7351

+++ reconf.sh
@@ -1,8 +1,23 @@
 #!/bin/bash
 #
+
+#VALG_EXEC=/home/cerion/Work/Valgrind/valgrind_trunk/Inst/bin/valgrind
+#VALG_SUPP=/home/cerion/Work/Valgrind/valgrind_trunk/Inst/lib/valgrind
+
+VALG_EXEC=/home/sewardj/Vg3LINE/trunk/Inst/bin/valgrind
+VALG_SUPP=/home/sewardj/Vg3LINE/trunk/Inst/lib/valgrind
+
+
+if [ ! -d "$QTDIR" ]; then
+    echo "error: QTDIR not set"
+    exit 1;
+fi
+
+OPTS="--shared --thread=yes --qt-dir=$QTDIR"
+
 bakClean -r
 rm -f ./bin/valkyrie
 make clean
-./configure --debug --thread=yes --vg-exec=/home/sewardj/Vg3LINE/trunk/Inst/bin/valgrind --vg-supp=/home/sewardj/Vg3LINE/trunk/Inst/lib/valgrind/
+./configure $OPTS --vg-exec=$VALG_EXEC --vg-supp=$VALG_SUPP
 #make
 #./bin/valkyrie
